#17 moved size stats check to a separate file
